requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
gspread
google-auth





